+++ rich/layout.py
@@ -191,6 +191,11 @@ class Layout:
         """Gets (visible) layout children."""
         return [child for child in self._children if child.visible]
 
+    @property
+    def map(self) -> RenderMap:
+        """Get a map of the last render."""
+        return self._render_map
+
     def get(self, name: str) -> Optional["Layout"]:
         """Get a named layout, or None if it doesn't exist.